Feat: add neat keybindings for fuzzy-finding
This commit adds key mappings for Telescope's built-in current buffer
fuzzy finder. The shortcut would be `<C-f>`.

Another one that was added was the ability to fuzzy find over the the
coniguration files. It uses the `vim.fn.stdpath('config')` function from
the Lua API to dynamically find the directory.